backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend con arquitectura hexagonal"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
	{ include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
motor = "^3.7.1"
argon2-cffi = "^25.1.0"
pyjwt = {version = ">=2.10.1", extras = ["crypto"]}

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2"
pytest-asyncio = ">=1.3.0"
black = ">=26.1.0"
mypy = "^1.8.0"
isort = ">=7.0.0"
ruff = ">=0.14.14"
radon = ">=5.2.0"
pytest-cov = ">=7.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=85"
markers = ["integration: tests that require a real MongoDB instance"]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
fail_under = 85

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "."
explicit_package_bases = true
